Deleted some models
In realization of my lack of skill as a 3d modeler (which was somehow
worse a year ago) I have deleted all of the player meshes and items in
the project.
Will add new meshes once I am capable of doing so. (correctly this time)
Also removed debugging logs from Structural Generation